Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 int main() {
     int r,area;
-    int pi = 3.14;
-    scanf("%d",&r);
+    float pi = 3.14;
+    scanf("%f",&r);
     area = pi*r*r;
-    printf("Area: %d",area);
+    printf("Area: %f",area);
     return 0;
 }
